Move gql queries out of components to their own module files for reusability
