--- assets/js/components/Project/queries.graphql
@@ -1,11 +1,0 @@
-query GetProjects {
-  projects {
-    id
-    title
-    folder
-    updated_at
-    ingestJobs {
-      id
-    }
-  }
-}
